unit test and fix for pdg rounding
The issue was that 10**(2-top_digit) would be a float if top_digit is > 2 and decimals (num) can't be multiplied by floats. However, if top_digit <= 2 then 10**(2-top_diigt) is an int and decimals CAN be multiplied by ints. The latter case is all that was tested in the test suite.